WIP: figure out how to do conditionals in egraphs
Summary of current approach:
* introduce a ternary 'select' node, which chooses between the second
  and third argument based on the value of the first argument
* At the end of a conditional statement, insert 'select' nodes to choose
  between values produced by each branch (analogous to phi, but
  controlled by the loop condition rather than the predecessor)
* In addition to selecting between explicit assignments of program
  variables, also select between IO-sequencing dummy values/tokens

The process of extracting a program with if-statements from an egraph
with select-statements is still murky, however.
The crux of the matter is that 'select' is basically an expression node,
but if-statements are *statements* and can only be put in certain
places. This requires introducing variables to assign, in possibly
complex ways.